ML-032/fix specs users request
