+++ spec/requests/api/v1/passwords_spec.rb
@@ -1,0 +1,34 @@
+require 'rails_helper'
+module Api
+  module V1
+    RSpec.describe 'Passwords', type: :request do
+      let(:user) { create :user }
+      let(:password_params) do
+        {
+          current_password: user.password,
+          password: 'newpassword',
+          password_confirmation: 'newpassword'
+        }
+      end
+
+      describe 'POST /create' do
+        it 'resets password' do
+          post '/api/v1/auth/password', params: {
+            email: user.email,
+            redirect_url: 'http://localhost:3000/reset-password'
+          }
+          expect(response).to have_http_status(:ok)
+        end
+      end
+
+      describe 'PUT /update' do
+        it 'updates password' do
+          put '/api/v1/auth/password', params: password_params,
+                                       headers: user.create_new_auth_token, as: :json
+          user.reload
+          expect(user).to be_valid_password('newpassword')
+        end
+      end
+    end
+  end
+end
